backend/requirements.txt
fastapi>=0.109.1
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
motor>=3.3.2
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.18
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-docx>=1.1.0
PyPDF2>=3.0.1
openai>=1.6.1
# Note: letta has known vulnerabilities in all versions <= 0.3.17
# No patched version available. Consider alternative personality management
# or implement additional access controls if using letta.
# letta>=0.3.0
rich>=13.7.0
watchdog>=3.0.0
python-dotenv>=1.0.0
